pyproject.toml
[project]
name = "bot-webhook"
version = "0.1.0"
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "flask>=1.0.3",
    "websockets>=4.0.1",
    "gunicorn>=20.1.0",
    "python-dotenv>=0.20.0",
    "pillow>=9.2.0",
    "requests>=2.28.1",
]
license = {text = "MIT"}
requires-python = ">=3.7"
[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
